src/main/java/com/student/student/controller/StudentController.java
package com.student.student.controller;

import com.student.student.exception.ResourceNotFoundException;
import com.student.student.model.Student;
import com.student.student.service.StudentService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/Student")
public class StudentController {

    @Autowired
    StudentService studentService;

    @GetMapping
    public ResponseEntity<List<Student>> getAllStudent() {
        return new ResponseEntity<>(studentService.getAllStudent(), HttpStatus.OK);
    }

    @PostMapping
    public ResponseEntity<Student> addStudent(@RequestBody Student student) {
        return new ResponseEntity<>(studentService.createStudent(student), HttpStatus.CREATED);
    }

    @GetMapping("{id}")
    public ResponseEntity<Student> getStudentById(@PathVariable String id) throws ResourceNotFoundException {
        return new ResponseEntity<>(studentService.getStudent(id), HttpStatus.OK);
    }

    @DeleteMapping("{id}")
    public ResponseEntity<String> deleteStudentById(@PathVariable String id){
        studentService.deleteStudent(id);
        return new ResponseEntity<String>("Student Successfully deleted",HttpStatus.OK);
    }

    @PutMapping("{id}")
    public ResponseEntity<Student> updateStudent(@PathVariable String id,@RequestBody Student newStudent){
        return new ResponseEntity<>(studentService.updateStudent(id, newStudent),HttpStatus.OK);
    }
}
